Services

+++ src/main/java/io/ims/backend/Repository/SubjectRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import io.ims.backend.Models.Subject;
 @Repository
-public interface SubjectRepository extends JpaRepository<Subject, Integer>{
+public interface SubjectRepository extends JpaRepository<Subject, Long>{
     
 }
